# docs/openapi/resource-center.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "new-api Resource Center API",
    "version": "2026-07-17",
    "summary": "Assets, asynchronous image tasks, uploads, and Webhooks.",
    "description": "Public Resource Center API. Async image tasks and uploads use a normal new-api token. Assets use an ak_ resource key. Account-level task Webhooks are configured in the Resource Center console."
  },
...
  "servers": [
    {
      "url": "{base_url}",
      "variables": {
        "base_url": {
          "default": "https://new-api.example.com",
          "description": "Your new-api origin."
        }
      }
    }
  ],
...
  "paths": {
...
    "/v1/image/tasks": {
      "post": {
        "tags": [
          "Async Images"
        ],
        "operationId": "createImageTask",
        "summary": "Create an asynchronous image task",
        "description": "Returns immediately after the task, billing reservation, credential lease, and durable dispatch are stored. Reusing an Idempotency-Key with the same normalized request returns the original task.",
        "security": [
          {
            "TokenAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Optional caller key, unique per user. Maximum 128 characters.",
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageTaskCreateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Task accepted.",
            "headers": {
              "Location": {
                "description": "Canonical task URL.",
                "schema": {
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "Suggested polling delay in seconds.",
                "schema": {
                  "type": "integer"
                }
              },
              "Idempotent-Replayed": {
                "description": "true when the original task is replayed.",
                "schema": {
                  "type": "boolean"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageTask"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "get": {
        "tags": [
          "Async Images"
        ],
        "operationId": "listImageTasks",
        "summary": "List asynchronous image tasks",
        "security": [
          {
            "TokenAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "Filter by one public task status.",
            "schema": {
              "$ref": "#/components/schemas/ImageTaskStatus"
            }
          },
          {
            "name": "operation",
            "in": "query",
            "required": false,
            "description": "Filter by generation or edit.",
            "schema": {
              "$ref": "#/components/schemas/ImageTaskOperation"
            }
          },
          {
            "name": "client_reference_id",
            "in": "query",
            "required": false,
            "description": "Filter by the caller business reference.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "created_after",
            "in": "query",
            "required": false,
            "description": "Exclusive lower creation-time bound in Unix seconds.",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "created_before",
            "in": "query",
            "required": false,
            "description": "Exclusive upper creation-time bound in Unix seconds.",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "description": "Task ID cursor returned by the previous page.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Page size.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cursor-paginated task list.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageTaskListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/v1/image/tasks/{task_id}": {
      "get": {
        "tags": [
          "Async Images"
        ],
        "operationId": "getImageTask",
        "summary": "Get an asynchronous image task",
        "security": [
          {
            "TokenAuth": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "description": "Public task ID such as task_xxx.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageTask"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/v1/image/uploads": {
      "post": {
        "tags": [
          "Image Uploads"
        ],
        "operationId": "uploadImageInputs",
        "summary": "Upload image inputs as multipart data",
        "description": "Accepts up to 10 repeated image files and one mask. Each PNG, JPEG, or WebP file is limited to 20 MiB and the request to 100 MiB. Temporary URLs should be submitted to a task promptly.",
        "security": [
          {
            "TokenAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ImageMultipartUploadRequest"
              },
              "encoding": {
                "image": {
                  "style": "form",
                  "explode": true
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Temporary uploaded URLs.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageUploadListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/v1/image/uploads/base64": {
      "post": {
        "tags": [
          "Image Uploads"
        ],
        "operationId": "uploadBase64ImageInputs",
        "summary": "Upload base64 image inputs",
        "description": "Accepts plain base64 or data URLs. Use images plus optional mask, or uploads with an optional field value of image or mask.",
        "security": [
          {
            "TokenAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageBase64UploadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Temporary uploaded URLs.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageUploadListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    }
  },
...
  "components": {
    "securitySchemes": {
      "TokenAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "new-api token",
        "description": "A normal new-api token authorized for the requested image model."
      },
      "AssetKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ak_*",
        "description": "A Resource Center key. Legacy keys have assets:read only."
      },
      "WebhookBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "wk-...",
        "description": "new-api generates this Key for the account-level task Webhook. The account owner can reveal, copy, or regenerate it in the Resource Center."
      }
    },
    "schemas": {
...
      "ImageTaskOperation": {
        "type": "string",
        "enum": [
          "generation",
          "edit"
        ]
      },
      "ImageTaskStatus": {
        "type": "string",
        "enum": [
          "queued",
          "in_progress",
          "succeeded",
          "failed"
        ]
      },
...
      "ImageTaskCreateRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "model",
          "operation",
          "input"
        ],
        "properties": {
          "model": {
            "type": "string",
            "minLength": 1
          },
          "operation": {
            "$ref": "#/components/schemas/ImageTaskOperation"
          },
          "input": {
            "$ref": "#/components/schemas/ImageTaskInput"
          },
          "output": {
            "$ref": "#/components/schemas/ImageTaskOutput"
          },
          "client_reference_id": {
            "type": "string",
            "maxLength": 191
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "allOf": [
          {
            "if": {
              "properties": {
                "operation": {
                  "const": "edit"
                }
              },
              "required": [
                "operation"
              ]
            },
            "then": {
              "properties": {
                "input": {
                  "type": "object",
                  "properties": {
                    "images": {
                      "type": "array",
                      "minItems": 1,
                      "items": {
                        "$ref": "#/components/schemas/ImageTaskSource"
                      }
                    }
                  },
                  "required": [
                    "images"
                  ]
                }
              }
            }
          }
        ]
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid credential.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Credential, scope, IP, expiry, or user policy denied access.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "The resource was not found for the authenticated user.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Conflict": {
        "description": "Idempotency conflict or endpoint limit.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Internal server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}